% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aggregate_k_nearest}
\alias{aggregate_k_nearest}
\title{Aggregate K nearest expression profiles}
\usage{
aggregate_k_nearest(Xr, Dr = matrix(), k = 2L, verbose = 0L)
}
\arguments{
\item{Xr}{A matrix in a shape of #genes x #samples.}

\item{Dr}{A predefined distance matrix in a shape of #samples x #samples. If
not specified, D is the distance matrix of the input \code{Xr}.}

\item{k}{An integer to choose \code{k} nearest samples (self-inclusive) to
aggregate based on the distance matrix \code{Dr}. If \code{k} is greater than
#samples, \code{k} is forced to be #samples to continue aggregation.}

\item{verbose}{An integer to specify verbose level.}
}
\value{
An aggregated matrix in a same shape of \code{Xr}.
}
\description{
Normalize (by median) and apply Freeman-Tukey transformation on the input
matrix \code{X}. Then calculate the distance matrix of samples.
}
