% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_euclidean}
\alias{dist_euclidean}
\title{Pair-wise euclidean distance among two matrices.}
\usage{
dist_euclidean(Ar, Br)
}
\arguments{
\item{Ar}{A matrix with samples on rows.}

\item{Br}{A matrix with samples on rows.}
}
\value{
A distance matrix in a shape of nrow(\code{Ar}) x nrow(\code{Br}).
}
\description{
Pair-wise euclidean distance among two matrices.
}
\examples{
x <- matrix(1:12, 3)
all.equal(c(as.matrix(dist(x))), c(dist_euclidean(x, x)))
}
\references{
\url{http://blog.felixriedel.com/2013/05/pairwise-distances-in-r/}
}
