\name{knnsmoother-package}
\alias{knnsmoother-package}
\alias{knnsmoother}
\docType{package}
\title{
\packageTitle{knnsmoother}
}
\description{
\packageDescription{knnsmoother}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knnsmoother}
\packageIndices{knnsmoother}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{knnsmoother}

Maintainer: \packageMaintainer{knnsmoother}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
