% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essence.R
\name{r_calculate_distances}
\alias{r_calculate_distances}
\title{Smoother performs normalization, transformation and calculates pair-wise distance}
\usage{
r_calculate_distances(mat)
}
\description{
Smoother performs normalization, transformation and calculates pair-wise distance
}
