% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essence.R
\name{r_knn_smoothing}
\alias{r_knn_smoothing}
\title{KNN-smoothing on UMI-filtered single-cell RNA-seq data}
\usage{
r_knn_smoothing(mat, k = 5)
}
\arguments{
\item{mat}{A numeric matrix with gene names on rows and cell names on columns.}

\item{k}{Number of nearest neighbours to aggregate.}
}
\value{
A smoothed numeric matrix.
}
\description{
KNN-smoothing on UMI-filtered single-cell RNA-seq data
}
\examples{
X <- matrix(abs(sin(seq(from=1, to=1000, length.out = 1000))),
nrow = 25, byrow = T)
y <- rep(1:4, each=10)
dim(X)
colnames(X) <- as.character(paste0("s", seq_len(ncol(X))))
rownames(X) <- as.character(paste0("g", seq_len(nrow(X))))
S <- knn_smoothing(X, k=5)
plot(X[1, ], X[3, ], col=factor(y), main="original")
plot(S[1, ], S[3, ], col=factor(y), main="smoothed")
}
