% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smoother_calc_distance}
\alias{smoother_calc_distance}
\title{Distance matrix}
\usage{
smoother_calc_distance(X, verbose = 0L)
}
\arguments{
\item{X}{A matrix in a shape of #genes x #samples.}

\item{verbose}{An integer to specify verbose level.}
}
\value{
A distance matrix in a shape of #samples x #samples.
}
\description{
Normalize (by median) and apply Freeman-Tukey transformation on the input
matrix \code{X}. Then calculate the distance matrix of samples.
}
